move the background steps into the another class to shared with injection
