+++ alert/src/test/java/org/opennms/horizon/alertservice/stepdefs/AlertTestSteps.java
@@ -35,7 +35,7 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.MessageOrBuilder;
 import com.google.protobuf.UInt64Value;
 import com.google.protobuf.util.JsonFormat;
-import io.cucumber.java.en.Given;
+
 import io.cucumber.java.en.Then;
 import io.restassured.RestAssured;
 import io.restassured.config.HttpClientConfig;
@@ -70,14 +70,7 @@ public class AlertTestSteps {
     private RetryUtils retryUtils;
     private KafkaTestHelper kafkaTestHelper;
     private AlertGrpcClientUtils clientUtils;
-
-    //
-    // Test Configuration
-    //
-    private String applicationBaseHttpUrl;
-    private String applicationBaseGrpcUrl;
-    private String kafkaBootstrapUrl;
-
+    private BackGroundSteps backGround;
 
     //
     // Test Runtime Data
@@ -88,57 +81,28 @@ public class AlertTestSteps {
     private Alert firstAlertFromLastResponse;
     private Long lastAlertId;
     private String testAlertReductionKey;
-    private String eventTopic;
-    private String alertTopic;
 
 //========================================
 // Constructor
 //----------------------------------------
 
-    public AlertTestSteps(RetryUtils retryUtils, KafkaTestHelper kafkaTestHelper, AlertGrpcClientUtils clientUtils) {
+    public AlertTestSteps(RetryUtils retryUtils, KafkaTestHelper kafkaTestHelper, AlertGrpcClientUtils clientUtils, BackGroundSteps bgSteps) {
         this.retryUtils = retryUtils;
         this.kafkaTestHelper = kafkaTestHelper;
         this.clientUtils = clientUtils;
+        this.backGround = bgSteps;
+        initKafka();
+    }
+
+    private void initKafka() {
+        kafkaTestHelper.setKafkaBootstrapUrl(backGround.getKafkaBootstrapUrl());
+        kafkaTestHelper.startConsumerAndProducer(backGround.getEventTopic(), backGround.getEventTopic());
+        kafkaTestHelper.startConsumerAndProducer(backGround.getAlertTopic(), backGround.getAlertTopic());
     }
 
 //========================================
 // Gherkin Rules
 //========================================
-
-    @Given("Kafka event topic {string}")
-    public void createKafkaTopicForEvents(String eventTopic) {
-        kafkaTestHelper.startConsumerAndProducer(eventTopic, eventTopic);
-        this.eventTopic = eventTopic;
-    }
-
-    @Given("Kafka alert topic {string}")
-    public void createKafkaTopicForAlerts(String alertTopic) {
-        kafkaTestHelper.startConsumerAndProducer(alertTopic, alertTopic);
-        this.alertTopic = alertTopic;
-    }
-
-    @Given("Application base HTTP URL in system property {string}")
-    public void applicationBaseHttpUrlInSystemProperty(String systemProperty) {
-        this.applicationBaseHttpUrl = System.getProperty(systemProperty);
-
-        log.info("Using base HTTP URL {}", this.applicationBaseHttpUrl);
-    }
-
-    @Given("Application base gRPC URL in system property {string}")
-    public void applicationBaseGrpcUrlInSystemProperty(String systemProperty) {
-        this.applicationBaseGrpcUrl = System.getProperty(systemProperty);
-
-        log.info("Using base gRPC URL: {}", this.applicationBaseGrpcUrl);
-    }
-
-    @Given("Kafka bootstrap URL in system property {string}")
-    public void kafkaRestServerURLInSystemProperty(String systemProperty) {
-        this.kafkaBootstrapUrl = System.getProperty(systemProperty);
-        this.kafkaTestHelper.setKafkaBootstrapUrl(kafkaBootstrapUrl);
-
-        log.info("Using Kafka base URL: {}", this.kafkaBootstrapUrl);
-    }
-
     @Then("Send event with UEI {string} with tenant {string}")
     public void sendMessageToKafkaAtTopic(String eventUei, String tenantId) {
         EventLog eventLog = EventLog.newBuilder()
@@ -151,7 +115,7 @@ public class AlertTestSteps {
                 .setUei(eventUei))
             .build();
 
-        kafkaTestHelper.sendToTopic(eventTopic, eventLog.toByteArray());
+        kafkaTestHelper.sendToTopic(backGround.getEventTopic(), eventLog.toByteArray());
     }
 
     @Then("List alerts for tenant {string}, with timeout {int}ms, until JSON response matches the following JSON path expressions")
@@ -209,7 +173,7 @@ public class AlertTestSteps {
     public void verifyTopicContainsTenant(int expectedMessages, String tenant) {
         int foundMessages = 0;
 
-        List<ConsumerRecord<String, byte[]>> records = kafkaTestHelper.getConsumedMessages(alertTopic);
+        List<ConsumerRecord<String, byte[]>> records = kafkaTestHelper.getConsumedMessages(backGround.getAlertTopic());
         for (ConsumerRecord<String, byte[]> record: records) {
             if (record.value() == null) {
                 continue;
@@ -266,7 +230,7 @@ public class AlertTestSteps {
     }
 
     private void commonSendGetRequestToApplication(String path) throws MalformedURLException {
-        URL requestUrl = new URL(new URL(this.applicationBaseHttpUrl), path);
+        URL requestUrl = new URL(new URL(backGround.getApplicationBaseHttpUrl()), path);
 
         RestAssuredConfig restAssuredConfig = this.createRestAssuredTestConfig();
